+++ Realm/ObjectServerTests/RLMSyncTestCase.h
@@ -40,7 +40,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 + (NSURL *)authServerURL;
 
-+ (RLMSyncCredential *)basicCredentialWithName:(NSString *)name createAccount:(BOOL)createAccount;
++ (RLMSyncCredential *)basicCredentialWithName:(NSString *)name register:(BOOL)shouldRegister;
 
 /// Synchronously open a synced Realm and wait until the binding process has completed or failed.
 - (RLMRealm *)openRealmForURL:(NSURL *)url user:(RLMSyncUser *)user;